[231-move-mark_actions-runtime-checks-to-decoration-time] - extract logic to a function
