+++ rapyer/base.py
@@ -1,13 +1,12 @@
 import base64
 import contextlib
 import functools
-import inspect
 import logging
 import pickle
 import uuid
 from collections.abc import AsyncIterator
 from contextlib import AbstractAsyncContextManager
-from typing import Any, ClassVar, Optional, cast, get_origin
+from typing import Any, ClassVar, Optional, get_origin
 
 from pydantic import (
     BaseModel,
@@ -26,11 +25,9 @@ from redis.commands.search.query import Query
 from redis.exceptions import NoScriptError, ResponseError
 
 from rapyer.actions import (
-    MARK_ACTION_PARAMS_ATTR,
     ActionGroup,
-    MarkActionParams,
     TargetSource,
-    _build_action_wrapper,
+    install_marked_action_methods,
     mark_actions,
     register_action_target,
     should_refresh_for_action,
@@ -296,50 +293,7 @@ class AtomicRedisModel(BaseModel):
         """
         This function is resposible for building the model according to the model configuration (For example, setting up ttl refresh actions)
         """
-        seen: set[str] = set()
-        for klass in cls.__mro__:
-            if klass is object:
-                break
-            for name, attr in vars(klass).items():
-                if name in seen:
-                    continue
-                if isinstance(attr, classmethod):
-                    raw_func = attr.__func__
-                    rebuild = classmethod
-                elif isinstance(attr, staticmethod):
-                    raw_func = attr.__func__
-                    rebuild = staticmethod
-                elif inspect.isfunction(attr):
-                    raw_func = attr
-                    rebuild = None
-                else:
-                    continue
-                seen.add(name)
-                params: MarkActionParams = cast(
-                    MarkActionParams, getattr(raw_func, MARK_ACTION_PARAMS_ATTR, None)
-                )
-                if params is None:
-                    continue
-                # If a parent subclass installed a wrapper, walk back to the
-                # truly-bare function so this subclass's decision starts fresh.
-                while hasattr(raw_func, "__wrapped__"):
-                    raw_func = raw_func.__wrapped__
-                should_refresh = (
-                    not params.ignore_refresh
-                    and inspect.iscoroutinefunction(raw_func)
-                    and should_refresh_for_action(cls.Meta, params.combined)
-                )
-                should_start_ttl = params.initial and cls.Meta.ttl
-                should_wrap = should_refresh or should_start_ttl
-                if should_wrap:
-                    installed = _build_action_wrapper(
-                        raw_func, params.combined, params.target, params.initial
-                    )
-                else:
-                    installed = raw_func
-                if rebuild is not None:
-                    installed = rebuild(installed)
-                setattr(cls, name, installed)
+        install_marked_action_methods(cls)
 
     def __init_subclass__(cls, **kwargs):
         # Find fields with KeyAnnotation and SafeLoadAnnotation
